requirements.txt
django
gunicorn
django-heroku
requests
djangorestframework
pandas
numpy
twilio